backend/requirements.txt
# Dependencies for Smart Ingredient Chef Backend
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
inference-sdk>=0.9.0
gunicorn>=21.0.0
